Option --find performs case insensitive name match
The current implementation of --find option is sub-optimal and will
certainly change in the future. It works properly only when exact match
is found. More precisely, even if case-insensitive match is correctly
identified, the output doesn't work properly.

Also notice that the actual implementation of match only finds exact
case-insensitive matches in name. There are several ways to perform
approximate match, some (like Dice comparison) have implementations on
the code. These implementations are private for the algorithm since
they are not used.